added float type for rating

+++ app/models/rating.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer
+from sqlalchemy import Column, ForeignKey, Integer, Float
 from sqlalchemy.orm import relationship
 
 from app.dependencies.database import Base
@@ -10,7 +10,7 @@ class Rating(Base):
     id = Column(Integer, primary_key=True, index=True)
     book_id = Column(Integer, ForeignKey("books.id", ondelete="CASCADE"), index=True)
     user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), index=True)
-    rating = Column(Integer, nullable=False)
+    rating = Column(Float, nullable=False)
 
     book = relationship("Book", back_populates="ratings")
     user = relationship("User", back_populates="ratings")
